[scrutiny][core] Tag components that come from static pkgs
This change makes the Scrutiny core plugin, which collects packages and
components, depend on the collector that to retrieves the static
packages list. This allows the core plugin to, when the static packages
list is available, tag components that come from a static package, allowing
controllers to verify that certain components come from trusted sources such as
ZBI and the static packages list.

Bug: 81537
Test: fx test scrutiny

Change-Id: I6d10118c2018ca136301ac1be8c7a5c453000b74
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/582541
Commit-Queue: Mark Dittmer <markdittmer@google.com>
Reviewed-by: Laura Peskin <pesk@google.com>
Reviewed-by: Benjamin Wright <benwright@google.com>