[component][fuzzing] Add Transceiver.
This CL adds the Transceiver class, which is used to convert Inputs to
and from Zircon sockets that can be exchanged with off-device code, e.g.
an ffx plugin via overnet. This class is responsible for transferring
file-like objects to and from the host. It uses a dedicated thread so as
not block either the FIDL dispatcher or runner when transferring
arbitrary amounts of data.

Bug: 24740
Change-Id: I2fde1730753e0b5be5b1090761e35dbbec838380
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/567109
Commit-Queue: Aaron Green <aarongreen@google.com>
Reviewed-by: Cameron Finucane <eep@google.com>