[virtualization][vmm] Decode OR instructions with opcode 80 and 81
Clang generates OR instructions with opcode 80 for virtio PCI backend
code. This type of instruction reads data from memory, performs an
inclusive or operation between the data and an immediate value, than
writes the result back to memory. This CL adds support for decoding
instructions with opcode 80 and 81, so that we can cover all cases where
an OR operations is performed between an immediate and a block of memory
with the same size.

Bug: 12611
Change-Id: I8d598c9e30041e7429c5d44929af8a9c6470e942
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/576731
Commit-Queue: Tony Liu <liutony@google.com>
Reviewed-by: Abdulla Kamar <abdulla@google.com>
Reviewed-by: David Greenaway <dgreenaway@google.com>