[fidlgen][llcpp] Simplify decl iteration
Our fidlgen templates have a lot of passages like:
{{- range .Decls }}
{{- if Eq .Kind Kinds.Const }}{{ template "ConstDeclaration" . }}{{- end }}
{{- if Eq .Kind Kinds.Protocol }}{{ template "ProtocolDeclaration" . }}{{- end }}
{{- if Eq .Kind Kinds.Struct }}{{ template "StructDeclaration" . }}{{- end }}
{{- if Eq .Kind Kinds.Table }}{{ template "TableDeclaration" . }}{{- end }}
{{- if Eq .Kind Kinds.Union }}{{ template "UnionDeclaration" . }}{{- end }}
{{- end }}
where we go through all of the declarations in the library and then for
some subset of declaration kinds we dispatch each to a different
template.

This CL introduces two new kinds of method on the root C++ IR object:
 - methods for each kind of declaration that return all of of that kind:
   .Consts, .Bits, .Enums, etc
 - .ProtocolForTransport that returns all protocols supporting a given
    transport.

All of the templates for llcpp have been rewritten to use these methods
instead of .Decls. The result is that the order of types generated has
changed but there's no semantic difference and I think it's more
readable.

Change-Id: I62e8431c3b0fd77204a3dfa51b2a20b46bb4882d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/587249
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Benjamin Prosnitz <bprosnitz@google.com>
Reviewed-by: Yifei Teng <yifeit@google.com>
Fuchsia-Auto-Submit: Ian McKellar <ianloic@google.com>