[rfc] Fixing RFC-0082 area
The `PosixCompat` was not listed in _areas.yaml, adding it there for
now. As noted in the TODO, we plan to align this list with overall
project areas, and can revisit this particular area choice at that point
in time. In the API world, we have the `fuchsia.posix` library prefix so
it seems sensible to have an API area dedicated to POSIX compliance.

Change-Id: Ib619bf34b0ac05a59374c98a6a27915f02df1f2f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/587403
Fuchsia-Auto-Submit: Pascal-Louis Perez <pascallouis@google.com>
Reviewed-by: Adam Barth <abarth@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>