[ffx] Rearrange self-test code
There was a question about adding tests to the self-test plugin. It
seemed like a bad idea to have a giant file that accumulates tests if
different plugins want to also put test inside there. I suggested
creating a module for each plugin to contain their own tests. I figured
it would be good to clean up the current tests to follow this pattern.
This should make it easier for people to add tests without merge
conflicts.

The only visible change is that the test macro stringifies the name to
the function to display so "test_daemon_echo" becomes
"daemon::test_echo" in the output. This shouldn't be too bad and it
actually helps organize the output a bit as you can see what the test is
roughly about first by the module name rather than having to read past
"test_" to get to it.

Change-Id: I4da1fd4081c0fbdbe3d068a89829a48e64fa593e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/584863
Reviewed-by: Jordon Wing <jwing@google.com>
Commit-Queue: Andy Weiss <dragonbear@google.com>