[flatland] Add ChildViewWatcher.GetViewRef() method.
This allows the embedder/parent to obtain the ViewRef that was
minted by the embeddee/child component.

This doesn't fully fix fxbug.dev/81850, because the view ref
is sent as soon as the connection is established; it doesn't
wait until content is first presented.

Bug: 81850
Change-Id: I3d2cf37e9e27342f11b1cb8aa8bb746713580ed0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/584304
Commit-Queue: Josh Gargus <jjosh@google.com>
Reviewed-by: Jaeheon Yi <jaeheon@google.com>
Reviewed-by: Emircan Uysaler <emircan@google.com>
API-Review: Jaeheon Yi <jaeheon@google.com>