[ffx] Add self-tests for config
This adds some simple env, get, and set tests for ffx config. As ffx config is
one of the top user flows it is desirable to have tests for it. This can
be fleshed out more in the future, but that will be easier with
something in place.

Bug: 82378

Change-Id: I52ebaf7b990fc5c9b2ab60ff005fcf2726ffd21a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/584864
Commit-Queue: Andy Weiss <dragonbear@google.com>
Reviewed-by: Matthew Boetger <boetger@google.com>