[kernel][vm] Flesh out aging and eviction threads.
This commit extends the existing simplistic aging thread to support
potential for multiple aging reasons, and LRU processing is moved out
of line into a separate thread.

There are no behavior changes in this commit, this just creates the
rest of the necessary framework for adding changes.

Bug: 84217
Bug: 83323
Change-Id: Ieeadf4a2be8d6146ea3fd975277e087b7bff57fc
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/573786
Reviewed-by: Rasha Eqbal <rashaeqbal@google.com>
Commit-Queue: Adrian Danis <adanis@google.com>