[codesize] mark library as not null safe
In order to allow our build rules to specify that a program
should compile with sound-null-safety we need to check if all the
libraries are marked as null-safe. This library is saying it is
null safe but opts out of all but 1 file which breaks the build
rule check. This CL will opt out of null safety but then opt in
for the one file that is using it.

Change-Id: I43e238a36a0c84df694a01332cf743a9469d6bbf
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/587792
Commit-Queue: Chase Latta <chaselatta@google.com>
Reviewed-by: Yifei Teng <yifeit@google.com>