[fshost] Format minfs when writing fails during a resize
Write a file to minfs to indicate when writing the contents of the old
instance to the new instance has started. This file is used to detect
if the device was restarted during the copy operation which could
potentially leave the device in a unknown state. If the file is detected
after a restart then minfs is reformatted.

Overwrite the minfs magic before removing the partition to prevent
mounting minfs on the next boot if deleting the partition partially
fails.

Bug: 84885
Change-Id: I80fd409671307586eb00c7fb6483435b1056231e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/586862
Reviewed-by: Brett Wilson <brettw@google.com>
Commit-Queue: Chris Drouillard <cdrllrd@google.com>