[fx] Make new-sync-to support download snapshots from GCS
In replacing the legacy internal `fx sync-to` command with a new
public version, we must continue to support syncing to every release
branch that's still under active development. That includes branches
from before recipes started emitting the `checkout_info` output
property that `fx new-sync-to` currently uses for determining how to
reproduce a checkout.

Therefore, if a build doesn't have the `checkout_info` output property
set, `fx new-sync-to` needs to replicate the legacy sync-to command's
behavior of downloading a Jiri snapshot that the build uploaded to GCS,
and syncing each repo to the revision pinned in that snapshot.

I tested this on some old release branches to make sure the
functionality works and is consistent with the legacy `fx sync-to`
command.

I also updated `fx repro` to suggest using `new-sync-to` for builds that
the legacy `sync-to` doesn't support.

Bug: 68492
Bug: 72533
Change-Id: I833035fd853f427b7cd29d8532ac8f5b0bf931d9
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/587621
Commit-Queue: Oliver Newman <olivernewman@google.com>
Fuchsia-Auto-Submit: Oliver Newman <olivernewman@google.com>
Reviewed-by: Gary Boone <gboone@google.com>