[mediaplayer] Fix audio renderer issues when used with AudioConsumer
AudioConsumer clients are allowed to start the timeline with SetRate
without specifying a subject time. This can lead to the AudioRenderer
being started on the incorrect media time, which causes
silence/stalls/glitches.

Add another state to FidlAudioRenderer to allow AudioConsumer to
defer calling Play till it's user has called Start with the real
subject time. Allow packets to be sent to the AudioRenderer in this
state to 'prime' it, and prevent stalling the upstream StreamSink.

Test: fx test mediaplayer_tests
Bug: 83451

Change-Id: I169ac066fa5b0a2f18711ca9ec6a126833fa59b1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/587646
Commit-Queue: Austin Foxley <afoxley@google.com>
Reviewed-by: Dale Sather <dalesat@google.com>