[virtualization][vsock] Simplify Mux/Demux control flow
The VirtioVsock::Mux and VirtioVsock::Demux methods are called by
a VirtioQueueWater object when descriptors become available on the RX/TX
virtio queues. While in general, the VirtioQueueWaiter is convenient for
virtio code (because it passes the index of the next free descriptor to
the callback), in the case of VirtioVsock it complicates the code
a little: for example, the Mux/Demux functions don't always want to use
a descriptor (and hence need to return it unused).

Replace VirtioQueueWaiter (and the associated wrapper class "Stream")
with just a plain async::WaitMethod object.

Multiply: virtualization-core-tests, vmm_unittests
Bug: fxbug.dev/85702

Change-Id: I761d28bc710304ffbd6e999e7ca46f0278748d4b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/584721
Commit-Queue: David Greenaway <dgreenaway@google.com>
Reviewed-by: Abdulla Kamar <abdulla@google.com>
Reviewed-by: Adrian Danis <adanis@google.com>