[events] Add EventSequence.has_subset
The has_subset sequence matcher allows matching against a subset
of events in the stream. Unlike all_of, the EventSequence will
not return an error if it sees an event that does not match
a matcher in the group.

This is useful if the event stream contains unrelated events, such
as those created by a different test within the same component.

Fixed: 85547
Change-Id: I5f1215f4d49d57510c15be2b956cfdb90f02b0ad
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/586603
Commit-Queue: Yegor Pomortsev <ypomortsev@google.com>
Reviewed-by: Gary Bressler <geb@google.com>
Reviewed-by: Fady Samuel <fsamuel@google.com>