[flatland] Wait for display image release signals in vulkan path
This CL solves "rendering into in-use backbuffer" problems for vulkan
rendering path in flatland. We exclusively wait for release signal before
presenting images to display. It is safe to do so in Vulkan path as we
are flipping between buffers.

This guards against vsync intervals longer than the hardcoded 14ms until
ApplyConfig2 is available.

Bug: 81179
Change-Id: I85e8e59b959cf047b362f05039ff1cba57c5d965
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/586872
Reviewed-by: Josh Gargus <jjosh@google.com>
Commit-Queue: Emircan Uysaler <emircan@google.com>