[ffx][repository] add-from-pm should default repo name to "devhost"
This incorporates feedback from UX that wants us to both standardize on our
syntax for specifying repository names, and to simplify the arguments in the
simple case where a user is adding a single repository. This switches
`ffx repository add-from-pm` from taking the repository name as a positional
argument, and over to having it be an argument `--repository $REPO_NAME`.
Furthermore, this changes the repository name to default to "default".

Fixed: 85199
Change-Id: Ia1983cbe0db4b744daf1c59486b76c431d581221
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/588191
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Jiaming Li <lijiaming@google.com>
Reviewed-by: Omer Lev-ran <omerlevran@google.com>
Fuchsia-Auto-Submit: Erick Tryzelaar <etryzelaar@google.com>