[bt][avrcp] Use PSM from search result for outgoing connections
It's possible for some devices to advertise a PSM for an AVRCP service
other than PSM_AVCTP. This is not likely, but the peer may expect us to
connect to it.
Update AVRCP to parse the PSM from the primary protocol of the BR/EDR
search result. Use this PSM to make outgoing connections.

Fixed: 63715
Test: fx test bt-avrcp-tests
Change-Id: Icad870765fc3677b943f3ed0e59dd25603ef84ff
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/581384
Fuchsia-Auto-Submit: Ani Ramakrishnan <aniramakri@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Marie Janssen 💖 <jamuraa@google.com>