[fidlc] Restrict @transitional to methods, @unknown to enum members
This CL removes a lot of complicated attribute validation around
@transitional and @unknown for enums, bits, and unions that was never
used for anything in backends. Now @transitional is only allowed on
methods (not on types as a way to allow @unknown despite strictness),
and @unknown is only allowed on enum members (not bits or unions).

Test: fx test fidl-compiler
Fixed: 84334

Change-Id: Id858d96e6b59f95cd7733aa897eda24fd67c9ba0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/587909
Commit-Queue: Mitchell Kember <mkember@google.com>
Reviewed-by: Pascal-Louis Perez <pascallouis@google.com>