[test_manager] Stop running a suite when the controller handle is dropped
Fixes a regression where the suite is no longer killed when it's
controller channel is closed. This could cause test manager to crash if
it attempts to send a suite event when the future serving the controller
is already finished.
This works because unlike an AbortHandle, a RemoteHandle cancels the
future when the handle is dropped.

Testing: fx test test_manager_test

Change-Id: Ib3c0b908e8a5c2fd06a221681f62511c79b58f30
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/586171
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Satsuki Ueno <satsukiu@google.com>
Reviewed-by: Ankur Mittal <anmittal@google.com>