[kernel][vm] Consider aspaces used if modified
Any operation that modifies the mapping in an arch aspace can modify it
in a similar way to what is being detected by active_since_last_check_
flag. Not setting the flag when these modifications happens will cause
scanning to falsely think the aspace has nothing new to scan, when it
might. This CL causes any operation that changes the aspace mappings to
treat the aspace as if it has been active, and so might have new
information in it.

Bug: 85512
Change-Id: I867e58610ddbe8aa91c6ab9c36b05a5285d156c4
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/587302
Reviewed-by: Rasha Eqbal <rashaeqbal@google.com>
Commit-Queue: Adrian Danis <adanis@google.com>