[fidl][rust] Add fidl::prelude
This CL adds a prelude module that re-exports common FIDL traits. It is
meant to be imported like `use fidl::prelude::*;`. This is similar to
futures::prelude which has widespread use.

There are two reasons for this:

1. It is convenient for bindings users.

2. Over time, we have been moving more bindings functionality into
   traits (e.g. see Idcad02ae7870d6a9fd26d6453cb49d467f353d31,
   https://fxbug.dev/85701). These moves are painful as they require
   adding imports to hundreds of files across the source tree. The
   prelude (if used everywhere) will solve that.

While glob imports are generally looked down upon, doing so for a
prelude (which typically contain only traits) is standard practice.

This also updates the Rust bindings docs.

Change-Id: Ia2bca686de937786763e3e7262ff796ec5d8f0be
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/587910
Commit-Queue: Mitchell Kember <mkember@google.com>
Reviewed-by: Tyler Mandry <tmandry@google.com>
Reviewed-by: Benjamin Prosnitz <bprosnitz@google.com>