[llcpp][driver] Minimal CL forking LLCPP templates for drivers
This CL forks LLCPP to create templates for drivers, and tries to do
so in a minimal way so that it can be done with minimal conflicts with
concurrent CLs. Because of this strategy, some changes like directory
reorganization (e.g. folder for drivers) are deferred until a later
point.

While it isn't ideal to have duplicate copies of a lot of the LLCPP
templates, drivers will need to modify most of these templates anyways
and it needs to be done at some point.

Bug: 85734
Change-Id: Iff4ebe16747314a55df5a14590b375e578b7f682