[llcpp] Add ContainsHandle trait for num handles calc
This CL adds a ContainsHandle trait which is true if the object contains
a handle. This is used for computing the number of handles when envelopes
are inlined. The previous approach used erroneously didn't count structs
with handles as containing handles.

Bug: 79578
Change-Id: I147d1915ad872790af42f365803ab461996abf5e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/588421
Reviewed-by: Pascal-Louis Perez <pascallouis@google.com>
Reviewed-by: Yifei Teng <yifeit@google.com>
API-Review: Pascal-Louis Perez <pascallouis@google.com>
Commit-Queue: Benjamin Prosnitz <bprosnitz@google.com>