[wlan] Fix bitfield/element conversion warnings
A lot of the wlan codebase is built with conversion warnings disabled.
When creating a new build target with conversion warnings enabled some
of the common files and one of the unittests came up with warnings as
errors. Fix these by casting to the proper types.

Test: fx test wlan_tests
Change-Id: Iacd783c214258a828d7983ff7b338fd5666bed97
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/588227
Reviewed-by: Charles Celerier <chcl@google.com>
Commit-Queue: Bjoern Johansson <bjoernj@google.com>