[subprocess] Wait for subprocess to finish after sending SIGTERM.
The subprocess command like botanist or testrunner may have cleanup
steps that it should run even if it gets a SIGTERM, such as cleaning up
temp directories or killing an ffx daemon. This change sends a SIGTERM
(instead of SIGKILL) to the subprocess and waits for it to handle the
signal properly before exiting. This is needed for using ffx in infra to
ensure the ffx daemon gets stopped properly even if testrunner gets
terminated early.

The logic is copied from
https://fuchsia.googlesource.com/infra/infra/+/d3b5c7a7a4bb0d4182cd57be72acd31462668e0f/cmd/catalyst/main.go#31.

Bug: 84869
Change-Id: I6e4f35d7bd999f34b22caa1d66bd78d04e75dff3
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/588281
Fuchsia-Auto-Submit: Ina Huh <ihuh@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Oliver Newman <olivernewman@google.com>