[dart] set HOME in update_3p_packages.py
A recent change to pub requires that the HOME variable be set
to cache some values. We do not want to rely on this variable
being set so we set it to a temporary directory for the duration
of our subprocess.

https: //github.com/dart-lang/pub/issues/3167
Change-Id: I1c86d3a84119bfdfa275c8e4690133d936fb5806
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/589164
Reviewed-by: Alexander Biggs <akbiggs@google.com>
Commit-Queue: Chase Latta <chaselatta@google.com>