[component][fuzzing] Add Dispatcher and improve Binding
This CL promotes fuzzing::FakeDispatcher to fuzzing::Dispatcher and
replaces usages of async_dispatcher_t pointers with shared pointers to
Dispatchers instead. This allows the async::Loop to be encapsulated, and
the thread it starts to be available to fuzzing::Binding. This allows
the latter to correctly determine when it is running on the dispatcher
thread and perform tasks directly, vs. when it should post tasks to the
dispatcher and wait.

Bug: 24740
Change-Id: I7f63636df596322445ebd85ae67ada750c877e85
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/585484
Commit-Queue: Aaron Green <aarongreen@google.com>
Reviewed-by: Cameron Finucane <eep@google.com>