[packet-formats] Better IPv6 protocol number logic
In `Ipv6PacketRaw`, store the body and protocol number together since,
if extension headers can be parsed, they are both present, and if not,
neither is present.

In `Ipv6PacketBuilder`, rename the `next_hdr` field to `proto` and
clarify that its value is the upper-layer protocol number, not the fixed
header's Next Header value (though they happen to be the same if no
extension headers are present).

Test: fx test packet-formats -- \
      test_packet_builder_proto_not_next_header

Change-Id: I58aa501156947ba66e344a2425c04ebf93a775c2
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/585083
Reviewed-by: Nick Brown <nickbrow@google.com>
Reviewed-by: Zeling Feng <zeling@google.com>
Commit-Queue: Joshua Liebow-Feeser <joshlf@google.com>