[utest][debugger] Allow the test's subprocesses to write to stdout/stderr
This is in preparation for converting these tests to using zxtest, so
that we can see the messages from assertion failures that occur in the
subprocesses.

I expect that these subprocesses had access to stdout/stderr when
these tests were originally written, but some later code changes
dropped that access.

Bug: 51652
Test: Tested by adding a printf() to the "argc >= 2" case in main.cc
Change-Id: I684831a38e6bcd1e5a864e61c1c2844aa382ba1e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/589006
Reviewed-by: Gianfranco Valentino <gevalentino@google.com>
Commit-Queue: Mark Seaborn <mseaborn@google.com>