[ffx][trace] Make list-providers fail better
In certain product configurations (e.g. userdebug)
fuchsia.tracing.controller.Controller is not available. In that scenario
you currently get a bad error message that just says PEER_CLOSED. This
looks for that error and instead provides a better error message.

I found that I also hit this if I did not have `//bundles:tools` in my
base image. So a very plain core.qemu-x64 build also hits this
PEER_CLOSED error. So the error message tries to help both of these
scenarios. I don't think we can do much better than that.

```
$ ffx trace list-providers
An attempt to access fuchsia.tracing.controller.Controller resulted in a bad status: PEER_CLOSED.
This can happen if tracing is not supported on the product configuration you are running or if it is missing from the base image.
```

Fixed: 84645

Change-Id: I2c0cd863af591a44dd2a4091cb7364f4e35f6dd2
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/588828
Reviewed-by: Andrew Davies <awdavies@google.com>
Commit-Queue: Andy Weiss <dragonbear@google.com>