[ffx] Ensure self-test cleans up daemons
This adds some code to the self-test runner to ensure that any daemons
are killed before finishing the tests. This is defensive and will cause
the test suite to fail if we did have to kill a daemon but should stop
the behavior in the attached bug.

Bug: 76646
Change-Id: Ide998e03f42177bd9bc12976d88d9773c9e3460e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/589405
Commit-Queue: Andy Weiss <dragonbear@google.com>
Reviewed-by: Andrew Davies <awdavies@google.com>