[audio_core][test] Free memory sooner during test
When RendererShim objects are created, an additional reference
is held until the test case exits. HermeticFidelityTest did not
realize this and created a new renderer for each of the 42
frequencies that it checks -- expecting the renderer and its
members to be freed automatically when the renderer went out of
scope.

HermeticFidelityTest now expressly calls Unbind on each
renderer, freeing memory immediately. This saves approx. 50MB
of memory, over the course of a HermeticFidelityTest case.

The problem was first noted -- and verified as fixed -- with
the log messages emitted by audio_core as it pins critical-path
memory (including renderer VMOs) every five seconds.

Change-Id: Ibd43bfe2c3912f02b5c7f207f721d937206dad8b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/588541
Reviewed-by: Tom Bergan <tombergan@google.com>
Reviewed-by: Austin Foxley <afoxley@google.com>
Commit-Queue: Martin Puryear <mpuryear@google.com>