[fx] Rename new-sync-to to sync-to
This new version of the command can be considered stable now that it's
passed all of my manual testing.

Internal checkouts will continue to use the internal version of the
command with the same name, but I'll soon deprecate it and rename it to
`fx old-sync-to` at which point all users will be switched to this new
version.

Bug: 72533
Change-Id: I050b4e69bcc7be8b926ee0817e50fec854e9297e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/589946
Commit-Queue: Oliver Newman <olivernewman@google.com>
Fuchsia-Auto-Submit: Oliver Newman <olivernewman@google.com>
Reviewed-by: Gary Boone <gboone@google.com>