[fshost] Shred zxcrypt on failed writes
Shred zxcrypt when writing the data back to minfs fails. This is done
to avoid putting components in unknown and untested states.

Change the return type of MaybeResizeMinfs so the caller knows
exactly what to do next.

Remove the CorruptMinfs step. Only the original superblock was being
overwritten and fsck would repair it from the backup superblock at the
start of the next boot. mkfs is unlikely to fail in a way that fsck
wouldn't detect and fsck failing will cause mkfs to run again.

Bug: 84885
Change-Id: I9c3227500822fe0299fec5c7e4a32421f2836cf5
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/588711
Commit-Queue: Chris Drouillard <cdrllrd@google.com>
Reviewed-by: James Sullivan <jfsulliv@google.com>
Reviewed-by: Brett Wilson <brettw@google.com>