[third_party][zstd] compile with ZSTD_MULTITHREAD when !is_fuchsia
We aim to enable multithreading in zstd when not compiling for fuchsia
(i.e., "host"); however, the current spelling of `is_host` fails to
cover the case of the linux toolchain. Recently, internal builds have
begun to build this library - in the context of the `zbi` tool - with
the linux toolchain and new errors have surfaced as a result. To better
reflect the original intent, `is_host` is replaced with  `!is_fuchsia`.
This change only mitigates the bug causing the error in the `zbi` tool;
a separate change will properly fix things there.

Bug: b/202053801
Change-Id: I2b9d3d311c688d86d9471d1658647b0dff841c6c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/590383
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Joshua Seaton <joshuaseaton@google.com>
Reviewed-by: Ashley Coleman <ashleycoleman@google.com>