[zbi][multithread] Check upper bound for workers
Checks the upper bound for the number of workers before setting it.
Zbi tool will fail when built without the ZSTD_MULTITHREAD flag. This
occurs because the max number of workers is 0 in single threaded
compression and the number of workers was set to hardware_concurrency()

Bug: b/202053801
Test: Built and ran ZBI tool without ZSTD_MULTITHREAD
Change-Id: I1cb05c08c49c173856062af52003a566f3e201c2
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/590387
Commit-Queue: Ashley Coleman <ashleycoleman@google.com>
Reviewed-by: Joshua Seaton <joshuaseaton@google.com>
Reviewed-by: Michael Brunson <mbrunson@google.com>