[utest][debugger] Change send_* and recv_* functions to use ZX_ASSERT
This is in preparation for converting these tests to using zxtest.

A subset of the send_*() and recv_*() functions are called from
inferior.cc in a subprocess (i.e. not the process where
zxtest/unittest runs), and so those functions can't be converted to
using zxtest's ASSERT/EXPECT macros for assertions.  Use ZX_ASSERT()
instead.  This requires removing uses of BEGIN_HELPER/END_HELPER.  For
some consistency, apply this to all of send_*() and recv_*().

All of the remaining functions in zircon/system/utest/debugger/ should
now be OK to switch from using unittest's assertions to zxtest's
assertions.

Bug: 51652
Test: fx set bringup.x64 --with zircon/system/utest/debugger/
  fx qemu -k
  /boot/test/debugger-test
Change-Id: I773d1b2adab7a0e67ef872ee575c77e04dc1fd9d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/590481
Reviewed-by: Gianfranco Valentino <gevalentino@google.com>
Commit-Queue: Mark Seaborn <mseaborn@google.com>