[hypervisor] Always cancel interrupts on kick
When kicking a VCPU, there is a possibility that it may be in the
process of executing a WFI or HLT instruction and have left the guest
but not currently be blocked on a WaitQueue. Because of this, we should
always attempt to cancel any pending wait-for-interrupts.

Manually verified this fix by running hypervisor-tests-pkg 1000 times on
a Sherlock with no failure of Guest.VcpuSuspendThread, which is
particularly good at picking up this problem.

Multiply: hypervisor-tests-pkg
Fixed: 86017
Change-Id: I0099b82289bb6551312c286f1e8df592fce1662b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/590328
Fuchsia-Auto-Submit: Abdulla Kamar <abdulla@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Adrian Danis <adanis@google.com>