[starnix] Fix infinite loop in sigsuspend
Sigsuspend would change the signal mask of the task, and then wait. Wait
would check the signal mask, see a pending signal, and return EINTR. The
signal mask would then be restored before the signal was actually
handled, and if the restored mask blocked the signal the signal would
not be handled.

This could cause clients to call sigsuspend, get an EINTR,
try again, forever.

This will also be useful when supporting the signal mask in epoll_pwait.

TEST: Updated sh_test to detect the failure

Change-Id: I829088d0f68a0b6fd222977527c1f177b301e849
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/590636
Commit-Queue: Kevin Lindkvist <lindkvist@google.com>
Reviewed-by: Theodore Dubois <tbodt@google.com>