[fshost] Keep the reboot reason cache file
The Copier only allows for excluding files when copying and it's a lot
easier to exclude all of the cache than to exclude everything except for
the reboot reason file. The reboot reason file is read separately and
inserted into the copier so it is included in the data size estimate and
written with the rest of the files.

Bug: 84885
Change-Id: I7ab19ce1284f80f04e499102d99d0896a14f3e79
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/589929
Commit-Queue: Chris Drouillard <cdrllrd@google.com>
Reviewed-by: James Sullivan <jfsulliv@google.com>
Reviewed-by: Brett Wilson <brettw@google.com>